fix a check that was always false
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2457 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ codeview.cpp
@@ -2265,7 +2265,7 @@ void CodeViewer::ReadLineToParseBrain(string wholeline, int linecount, vector<Us
 
 void CodeViewer::RemoveByVal(string &line)
 {
-	const size_t LL = line.length();
+	const int LL = line.length();
 	string SearhLine = lowerCase(line);
 	int Pos = SureFind( SearhLine, "byval");
 	if (Pos > -1)
